+++ src/stubServer.ts
@@ -15,15 +15,57 @@ type Stub = URL | StubFilename | GetStubFilenameFunction;
 
 type Delay = { min: number; max: number };
 
-type CommonConfig = { delay?: Delay; headers?: IncomingHttpHeaders };
+type CommonConfig = {
+  /**
+   * Delay in ms before processing the request.
+   *
+   * Simulates an unreliable network connection.
+   * Help find bugs and possible improvements (add a spinner, disable a submit button...) to your code.
+   */
+  delay?: Delay;
+
+  /**
+   * HTTP headers of the request received by the stub server to override.
+   */
+  headers?: IncomingHttpHeaders;
+};
 
 type Response = {
-  [requestMethod in RequestMethod]?: Stub | (CommonConfig & { response: Stub });
+  [requestMethod in RequestMethod]?:
+    | Stub
+    | (CommonConfig & {
+        /**
+         * The response the stub server will return.
+         *
+         * Can be:
+         *
+         * - a file:
+         *
+         *   - .json, .html, .jpg, .txt...: will be returned as is,
+         *     the returned HTTP status code is determined by the file name (ex: `my-api_400_BadRequest.json`)
+         *
+         *   - .js, .ts:
+         *     - can default export an [Express handler function](http://expressjs.com/en/guide/routing.html#route-handlers) with req and res
+         *     - otherwise will be returned as is (useful when you need a .json file with comments)
+         *
+         * - a URL: the request will be routed there (proxy mode)
+         *
+         * - a callback with [req](http://expressjs.com/en/4x/api.html#req) as parameter,
+         *   should return a file name or a URL
+         */
+        response: Stub;
+      });
 };
 
 type Route = CommonConfig & Response;
 
+/**
+ * Stub server config.
+ */
 export type StubServerConfig = CommonConfig & {
+  /**
+   * List of HTTP routes to handle.
+   */
   routes: {
     [apiPath: string]: Route;
   };
@@ -60,6 +102,8 @@ function getConfig() {
   return require(_configPath).default as StubServerConfig;
 }
 
+let _delay: boolean;
+
 async function parseConfig(apiPath: string, req: express.Request) {
   // Re-read the config file for each new request so the user
   // don't have to restart the stub server
@@ -95,7 +139,7 @@ async function parseConfig(apiPath: string, req: express.Request) {
   // Delay the request to simulate network latency
   // istanbul ignore next
   const { min, max } = delay ?? routeDelay ?? globalDelay ?? { min: 0, max: 0 };
-  const actualDelay = await randomDelay(min, max);
+  const actualDelay = _delay ? await randomDelay(min, max) : 0;
 
   console.log(`${requestMethod} ${req.url} => ${name}, delay: ${actualDelay} ms`);
 
@@ -130,6 +174,7 @@ async function processStubRequest(
         if (['.json', '.js', '.ts'].includes(path.extname(filename))) {
           // If file does not exist: "Cannot find module '...'"
           deleteRequireCache(filename);
+          // eslint-disable-next-line unicorn/no-await-expression-member
           const jsFileContent = (await import(filename)).default as object | ExpressHandlerFunction;
 
           if (typeof jsFileContent === 'function') {
@@ -162,11 +207,26 @@ async function processStubRequest(
 // https://expressjs.com/en/4x/api.html#router.METHOD
 type ExpressMethod = 'get' | 'post' | 'put' | 'patch' | 'delete' | 'options' | 'head';
 
-export function stubServer(configPath: string, app: express.Application) {
+type Options = {
+  /**
+   * If false, ignore any delay specified in the config.
+   */
+  delay?: boolean;
+};
+
+/**
+ * Starts the stub server: responds to HTTP requests based on the config.
+ *
+ * @param configPath path to the config file
+ * @param app the [Express application](https://expressjs.com/en/4x/api.html#express)
+ * @param options optional parameters
+ */
+export function stubServer(configPath: string, app: express.Application, options: Options = {}) {
   // Do not use asynchronous code here otherwise routes will
   // be defined after the ones from webpack-dev-server
 
   _configPath = configPath;
+  _delay = options.delay ?? true;
 
   const { routes } = getConfig();